PSR-1389 Moved upscan callback route to prod.routes and passed in internal URL to upscan
