+++ conf/app.routes
@@ -20,8 +20,6 @@
 
 GET         /                                                                                                  controllers.IndexController.onPageLoad()
 
-POST        /fileupload/upscan-callback                                                                        controllers.UploadCallbackController.callback
-
 GET         /assets/*file                                                                                      controllers.Assets.versioned(path="/public", file: Asset)
 
 GET         /language/:lang                                                                                    controllers.LanguageSwitchController.switchToLanguage(lang: String)
