+++ conf/application.conf
@@ -84,6 +84,12 @@ microservice {
           port = 10700
       }
 
+      pensionSchemeReturnFrontend {
+          protocol = http
+          host = localhost
+          port = 10701
+      }
+
       email {
           protocol = http
           host = localhost
@@ -152,6 +158,7 @@ urls {
   sippContinueJourney       = "/what-you-will-need"
   sippViewAndChange         = "/view-change-question"
   emailCallback             = "/pension-scheme-return/psr/email-response/%s/%s/%s/%s/%s/%s/%s/%s/%s"
+  upscanCallback            = "/fileupload/upscan-callback"
 
   manage-pension-schemes {
     baseUrl  = "http://localhost:8204/manage-pension-schemes"
